stopping criteria inherit from base class
