+++ src/probnum/quad/solvers/__init__.py
