--- src/probnum/quad/stop_criteria/__init__.py
@@ -1,6 +1,0 @@
-from ._stopping_criterion import (
-    IntegralVarianceTolerance,
-    MaxNevals,
-    RelativeMeanChange,
-    StoppingCriterion,
-)
